s3 push first of last person then take new pics so alexa knows who actually took last cup

+++ coffee.py
@@ -1,4 +1,5 @@
 import RPi.GPIO as GPIO
+import os
 import time
 import subprocess
 import json
@@ -9,6 +10,7 @@ from watson_developer_cloud import VisualRecognitionV3
 
 # config setup
 last_image_loc = join(dirname(__file__), 'last.jpg')
+last_name_loc = join(dirname(__file__), 'last.txt')
 config = {}
 with open(join(dirname(__file__), 'config.json'), 'rb') as data:
   config.update(json.load(data))
@@ -48,32 +50,34 @@ def watson_check():
     accuracy = str(persons[0]["score"])
     print('Identified ' + accused + ' (' + accuracy + ')')
     
-    return accused
+    # save accused name
+    last_name = open(last_name_loc, "w")
+    last_name.write(accused)
+    last_name.close()
     
-def upload_to_aws(accused):
-  # upload last image to S3
-  print('Uploading pic to S3')
-  last_pic = open(last_image_loc, 'rb')
-  s3.Bucket(config["s3_bucket"]).put_object(Key='last.jpg', Body=last_pic)
+def upload_to_aws():
+  if os.path.isfile(last_image_loc) and os.path.isfile(last_name_loc):
+    # upload last image to S3
+    print('Uploading pic to S3')
+    last_pic = open(last_image_loc, 'rb')
+    s3.Bucket(config["s3_bucket"]).put_object(Key='last.jpg', Body=last_pic)
 
-  # upload last name to S3
-  print('Uploading name to S3')
-  last_name = open("last.txt", "w")
-  last_name.write(accused)
-  last_name.close()
-  last_name = open("last.txt", 'rb')
-  s3.Bucket(config["s3_bucket"]).put_object(Key='last.txt', Body=last_name)
+    # upload last name to S3
+    print('Uploading name to S3')
+    last_name = open(last_name_loc, 'rb')
+    s3.Bucket(config["s3_bucket"]).put_object(Key='last.txt', Body=last_name)
   
 while True:
   input_state = GPIO.input(18)
   if input_state == False:
     print('Button Pressed')
+    
+    upload_to_aws()
+    
     take_pic()
     
-    accused = watson_check()
+    watson_check()
     #watson_check_debug()
-    
-    upload_to_aws(accused)
     
     print('Waiting 10 seconds')
     time.sleep(10)
